moving filter components into filter dir #238
